% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_add_item.R
\name{FFQ.add.fruit}
\alias{FFQ.add.fruit}
\title{Add food items to reference list}
\usage{
FFQ.add.fruit(
  FFQobject,
  Fooditem,
  Average_weight,
  Amount_of_fibre_g.100g,
  SDF_Hemicellulose = 0,
  SDF_Pectin = 0,
  ISD_Hemicellulose = 0,
  ISD_Cellulose = 0,
  ISD_Pectin = 0,
  ISD_Lignin = 0,
  RS1 = 0,
  RS2 = 0,
  RS3 = 0,
  Chitin = 0,
  Fructan = 0
)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}
}
\value{
Add food items to the reference lists in the FFQ object.
}
\description{
Add food items to reference list
}
\examples{
FFQobject <- FFQ.add.fruit(FFQobject = FFQobject, Fooditem = "Avocado", Average_weight = 170, Amount_of_fibre_g.100g = 6.7, SDF_Hemicellulose =0.5, SDF_Pectin=0.8, ISD_Hemicellulose=0.9, ISD_Cellulose=1.4, ISD_Pectin=0.2, ISD_Lignin=0,1, RS1=0, RS2=0, RS3=0, Chitin=0, Fructan=0)
}
