% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_calculate_bread_fibres.R
\name{FFQ.calculate.bread.fibres}
\alias{FFQ.calculate.bread.fibres}
\title{Calculate the fibre consumptions of bread}
\usage{
FFQ.calculate.bread.fibres(FFQobject, ignore.groups = FALSE)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}

\item{ignore.groups}{Check if FFQ food items are present in reference list. Default = FALSE.}
}
\value{
Calculate the fibre consumptions from bread out of the FFQ object.
}
\description{
Calculate the fibre consumptions of bread
}
\examples{
FFQobject_bread_fibres <- FFQ.calculate.bread.fibres(FFQobject)
}
\seealso{
CMET.FFQ::FFQ.calculate.fibres() .
}
