% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_calculate_fibres.R
\name{FFQ.calculate.fibres}
\alias{FFQ.calculate.fibres}
\title{Calculate the fibre consumptions from all food groups}
\usage{
FFQ.calculate.fibres(FFQobject, ignore.groups = FALSE)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}

\item{ignore.groups}{Check if FFQ food items are present in reference list. Default = FALSE.}
}
\value{
Calculate the fibre consumptions from all food groups out of the FFQ object.
}
\description{
Calculate the fibre consumptions from all food groups
}
\examples{
FFQobject_all_fibres <- FFQ.calculate.fibres(FFQobject)
}
