% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_convert_dataframe.R
\name{FFQ.convert.dataframe}
\alias{FFQ.convert.dataframe}
\title{Convert data frame into FFQ object}
\usage{
FFQ.convert.dataframe(
  dataframe,
  Object = "FFQ",
  datalist = "internal",
  datalist2 = datalist
)
}
\arguments{
\item{Object}{Transform .}

\item{FFQ_dataframe}{Dataframe exported from QuestionPro}
}
\value{
A custom object that contains everything required for downstream processing
}
\description{
Convert data frame into FFQ object
}
\examples{
FFQobject <- FFQ.convert.dataframe(FFQ_dataframe)
}
