% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_plot_fibres.R
\name{FFQ.plot.fibregroups}
\alias{FFQ.plot.fibregroups}
\title{Plot where the amount of fibres per food group}
\usage{
FFQ.plot.fibregroups(FFQobject)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}
}
\value{
Plot where the amount of fibres per food group.
}
\description{
Plot where the amount of fibres per food group
}
\examples{
plot <- FFQ.plot.fibregroups(FFQobject)
}
