% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_print_results.R
\name{FFQ.print.medium}
\alias{FFQ.print.medium}
\title{Print the medium}
\usage{
FFQ.print.medium(
  FFQobject,
  show.as.plain.text = FALSE,
  show.seeds = FALSE,
  divide = TRUE
)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}
}
\value{
Prints out the I-SHIME medium separately as a data frame. Easier to copy like this!
}
\description{
Print the medium
}
\examples{
FFQ.print.medium(FFQobject)
}
