% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_print_results.R
\name{FFQ.print.protocol}
\alias{FFQ.print.protocol}
\title{Print the I-SHIME protocol}
\usage{
FFQ.print.protocol(
  FFQobject,
  show.as.plain.text = FALSE,
  show.seeds = FALSE,
  divide = TRUE
)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}

\item{show.as.plain.text}{show the results as plain text}

\item{show.seeds}{Show and incorporate seeds}
}
\value{
Prints out the I-SHIME protocol.
}
\description{
Print the I-SHIME protocol
}
\examples{
FFQ.print.protocol(FFQobject)
}
