% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_print_results.R
\name{FFQ.print.transit}
\alias{FFQ.print.transit}
\title{Print the transit time}
\usage{
FFQ.print.transit(FFQobject)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}
}
\value{
Prints out the transit time separately as a data frame. Easier to copy like this!
}
\description{
Print the transit time
}
\examples{
FFQ.print.transit(FFQobject)
}
