% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_replace_reference.R
\name{FFQ.replace.reference}
\alias{FFQ.replace.reference}
\title{Replace refences}
\usage{
FFQ.replace.reference(FFQobject, New.reference, Reference.type)
}
\arguments{
\item{FFQobject}{FFQ object with all FFQ data}

\item{New.reference}{a dataframe of the new reference}

\item{Reference.type}{The type of reference. Possibilities are: "fruit", "vegetable", "nut" or "grain".}
}
\value{
Replace fibre reference in FFQ object.
}
\description{
Replace refences
}
\examples{
FFQobject_updated <- FFQ.replace.reference(FFQobject, dataframe_fruits, "fruit")
}
