% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFQ_classes.R
\docType{class}
\name{FFQobject-class}
\alias{FFQobject-class}
\title{The main experiment-level class for FFQobject data
In the case of missing component data, the slots are set to \code{NULL}. As
soon as a \code{FFQobject-class} object is to be updated with new component
data (previously missing/\code{NULL} or not), the indices of all components
are re-checked for compatibility and trimmed if necessary. This is to ensure
by design that components describe the same taxa/samples, and also that these
trimming/validity checks do not need to be repeated in downstream analyses.}
\description{
The main experiment-level class for FFQobject data
In the case of missing component data, the slots are set to \code{NULL}. As
soon as a \code{FFQobject-class} object is to be updated with new component
data (previously missing/\code{NULL} or not), the indices of all components
are re-checked for compatibility and trimmed if necessary. This is to ensure
by design that components describe the same taxa/samples, and also that these
trimming/validity checks do not need to be repeated in downstream analyses.
}
